Unify OS version xunit attributes
